code refactor

+++ lib/CustomHID/keymap.c
@@ -3,7 +3,7 @@
 // ---------------------------
 // Keymaps definition
 // ---------------------------
-uint16_t keymaps[MAX_LAYERS][MATRIX_ROWS][MATRIX_COLS] = {
+uint32_t keymaps[MAX_LAYERS][MATRIX_ROWS][MATRIX_COLS] = {
     // --- Layer 0: Base ---
     {
         {
